requirements.txt
PyQt6
notion-client==2.2.1
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0 
pyyaml